requirement.txt
Babel==2.8.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pycparser==2.20
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
